+++ src/paths/widest_paths.c
@@ -36,26 +36,29 @@
  * \function igraph_get_widest_paths
  * \brief Widest paths from a single vertex.
  *
- * Calculates the widest paths from a single node to all other specified nodes,
- * using a modified Dijkstra's algorithm.  If there is more than one path with
- * the largest width between two vertices, this function gives only one of them.
+ * Calculates the widest paths from a single vertex to all other specified
+ * vertices, using a modified Dijkstra's algorithm. The width of a path is
+ * defined as the width of the narrowest edge in the path.If there is more than
+ * one path with the largest width between two vertices, this function gives
+ * only one of them.
+ *
  * \param graph The graph object.
- * \param vertices The result, the IDs of the vertices along the paths.
- *        This is a list of integer vectors where each element is an
- *        \ref igraph_vector_int_t object. The list will be resized as needed.
- *        Supply a null pointer here if you don't need these vectors.
- * \param edges The result, the IDs of the edges along the paths.
- *        This is a list of integer vectors where each element is an
- *        \ref igraph_vector_int_t object. The list will be resized as needed.
- *        Supply a null pointer here if you don't need these vectors.
- * \param from The id of the vertex from/to which the widest paths are
- *        calculated.
+ * \param vertices The result, the IDs of the vertices along the paths. This
+ *    is a list of integer vectors where each element is an
+ *    \ref igraph_vector_int_t object. The list will be resized as needed.
+ *    Supply a null pointer here if you don't need these vectors.
+ * \param edges The result, the IDs of the edges along the paths. This is a
+ *    list of integer vectors where each element is an
+ *    \ref igraph_vector_int_t object. The vector list will be resized as
+ *    needed. Supply a null pointer here if you don't need these vectors.
+ * \param from The ID of the vertex from/to which the widest paths are
+ *    calculated.
  * \param to Vertex sequence with the IDs of the vertices to/from which the
- *        widest paths will be calculated. A vertex might be given multiple
- *        times.
- * \param weights The edge weights. Edge weights can be negative. If this
- *        is a null pointer or if any edge weight is NaN, then an error
- *        is returned. Edges with positive infinite weight are ignored.
+ *    widest paths will be calculated. A vertex may be given multiple times.
+ * \param weights The edge weights, interpreted as widths. Edge weights can be
+ *    negative, but must not be NaN. Edges with negative infinite weight are
+ *    ignored. The weight vector is required: if \c NULL is passed, an error is
+ *    raised.
  * \param mode The type of widest paths to be used for the
  *        calculation in directed graphs. Possible values:
  *        \clist
@@ -67,23 +70,20 @@
  *          the directed graph is considered as an
  *          undirected one for the computation.
  *        \endclist
- * \param parents A pointer to an initialized igraph vector or null.
- *        If not null, a vector containing the parent of each vertex in
- *        the single source widest path tree is returned here. The
- *        parent of vertex i in the tree is the vertex from which vertex i
- *        was reached. The parent of the start vertex (in the \c from
- *        argument) is -1. If the parent is -2, it means
- *        that the given vertex was not reached from the source during the
- *        search. Note that the search terminates if all the vertices in
- *        \c to are reached.
- * \param inbound_edges A pointer to an initialized igraph vector or null.
- *        If not null, a vector containing the inbound edge of each vertex in
- *        the single source widest path tree is returned here. The
- *        inbound edge of vertex i in the tree is the edge via which vertex i
- *        was reached. The start vertex and vertices that were not reached
- *        during the search will have -1 in the corresponding entry of the
- *        vector. Note that the search terminates if all the vertices in
- *        \c to are reached.
+ * \param parents A pointer to an initialized igraph vector or null. If not
+ *    null, a vector containing the parent of each vertex in the single source
+ *    widest path tree is returned here. The parent of vertex \c i in the tree
+ *    is the vertex from which vertex \c i was reached. The parent of the start
+ *    vertex (in the \p from argument) is -1. If the parent is -2, it means
+ *    that the given vertex was not reached from the source during the search.
+ *    The search terminates when all the vertices in \p to have been reached.
+ * \param inbound_edges A pointer to an initialized igraph vector or \c NULL.
+ *    If not \c NULL, a vector containing the inbound edge of each vertex in
+ *    the single source widest path tree is returned here. The inbound edge of
+ *    vertex \c i in the tree is the edge via which vertex \c i was reached.
+ *    The start vertex and vertices that were not reached during the search
+ *    will have -1 in the corresponding entry of the vector. The search
+ *    terminates when all the vertices in \p to have been reached.
  * \return Error code:
  *        \clist
  *        \cli IGRAPH_ENOMEM
@@ -113,21 +113,21 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
 
     /* Implementation details: This is a Dijkstra algorithm with a
     binary heap, modified to support widest paths. The heap is indexed,
-    so it stores both the widest path to a node, as well as it's index. We
+    so it stores both the widest path to a vertex, as well as it's index. We
     use a 2 way heap so that we can query indexes directly in the heap.
 
     To adapt a Dijkstra to handle widest path, instead of prioritising candidate
-    nodes with the minimum distance, we prioritise those with the maximum
-    width instead. When adding a node into our set of 'completed' nodes, we
-    update all neighbouring nodes with a width that is equal to the min of the
-    width to the current node and the width of the edge.
+    vertices with the minimum distance, we prioritise those with the maximum
+    width instead. When adding a vertex into our set of 'completed' vertices, we
+    update all neighbouring vertices with a width that is equal to the min of the
+    width to the current vertex and the width of the edge.
 
-    We denote the widest path from a node to itself as infinity, and the widest
-    path from a node to a node it cannot reach as negative infinity.
+    We denote the widest path from a vertex to itself as infinity, and the widest
+    path from a vertex to a vertex it cannot reach as negative infinity.
     */
 
-    igraph_integer_t no_of_nodes = igraph_vcount(graph);
-    igraph_integer_t no_of_edges = igraph_ecount(graph);
+    const igraph_integer_t vcount = igraph_vcount(graph);
+    const igraph_integer_t ecount = igraph_ecount(graph);
     igraph_vit_t vit;
     igraph_2wheap_t Q;
     igraph_lazy_inclist_t inclist;
@@ -140,10 +140,10 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
         IGRAPH_ERROR("Weight vector is required.", IGRAPH_EINVAL);
     }
 
-    if (igraph_vector_size(weights) != no_of_edges) {
+    if (igraph_vector_size(weights) != ecount) {
         IGRAPH_ERRORF("Weight vector length (%" IGRAPH_PRId ") does not match number of edges (%" IGRAPH_PRId ").",
                       IGRAPH_EINVAL,
-                      igraph_vector_size(weights), no_of_edges);
+                      igraph_vector_size(weights), ecount);
     }
 
     if (igraph_vector_is_any_nan(weights)) {
@@ -160,19 +160,19 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
         IGRAPH_CHECK(igraph_vector_int_list_resize(edges, IGRAPH_VIT_SIZE(vit)));
     }
 
-    IGRAPH_CHECK(igraph_2wheap_init(&Q, no_of_nodes));
+    IGRAPH_CHECK(igraph_2wheap_init(&Q, vcount));
     IGRAPH_FINALLY(igraph_2wheap_destroy, &Q);
     IGRAPH_CHECK(igraph_lazy_inclist_init(graph, &inclist, mode, IGRAPH_LOOPS));
     IGRAPH_FINALLY(igraph_lazy_inclist_destroy, &inclist);
 
-    IGRAPH_VECTOR_INIT_FINALLY(&widths, no_of_nodes);
+    IGRAPH_VECTOR_INIT_FINALLY(&widths, vcount);
     igraph_vector_fill(&widths, -IGRAPH_INFINITY);
 
-    parent_eids = IGRAPH_CALLOC(no_of_nodes, igraph_integer_t);
+    parent_eids = IGRAPH_CALLOC(vcount, igraph_integer_t);
     IGRAPH_CHECK_OOM(parent_eids, "Insufficient memory for widest paths.");
     IGRAPH_FINALLY(igraph_free, parent_eids);
 
-    is_target = IGRAPH_CALLOC(no_of_nodes, bool);
+    is_target = IGRAPH_CALLOC(vcount, bool);
     IGRAPH_CHECK_OOM(is_target, "Insufficient memory for widest paths.");
     IGRAPH_FINALLY(igraph_free, is_target);
 
@@ -182,7 +182,7 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
         if (!is_target[ IGRAPH_VIT_GET(vit) ]) {
             is_target[ IGRAPH_VIT_GET(vit) ] = true;
         } else {
-            to_reach--;       /* this node was given multiple times */
+            to_reach--;       /* this vertex was given multiple times */
         }
     }
 
@@ -212,8 +212,8 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
             igraph_real_t edgewidth = VECTOR(*weights)[edge];
             igraph_real_t altwidth = maxwidth < edgewidth ? maxwidth : edgewidth;
             igraph_real_t curwidth = VECTOR(widths)[tto];
-            if (edgewidth == IGRAPH_INFINITY) {
-                /* Ignore edges with infinite weight */
+            if (edgewidth == -IGRAPH_INFINITY) {
+                /* Ignore edges with negative infinite weight */
             } else if (curwidth < 0) {
                 /* This is the first assigning a width to this vertex */
                 VECTOR(widths)[tto] = altwidth;
@@ -235,9 +235,9 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
 
     /* Create `parents' if needed */
     if (parents) {
-        IGRAPH_CHECK(igraph_vector_int_resize(parents, no_of_nodes));
+        IGRAPH_CHECK(igraph_vector_int_resize(parents, vcount));
 
-        for (i = 0; i < no_of_nodes; i++) {
+        for (i = 0; i < vcount; i++) {
             if (i == from) {
                 /* i is the start vertex */
                 VECTOR(*parents)[i] = -1;
@@ -253,9 +253,9 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
 
     /* Create `inbound_edges' if needed */
     if (inbound_edges) {
-        IGRAPH_CHECK(igraph_vector_int_resize(inbound_edges, no_of_nodes));
+        IGRAPH_CHECK(igraph_vector_int_resize(inbound_edges, vcount));
 
-        for (i = 0; i < no_of_nodes; i++) {
+        for (i = 0; i < vcount; i++) {
             if (parent_eids[i] <= 0) {
                 /* i was not reached */
                 VECTOR(*inbound_edges)[i] = -1;
@@ -268,7 +268,7 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
     /* Reconstruct the widest paths based on vertex and/or edge IDs */
     if (vertices || edges) {
         for (IGRAPH_VIT_RESET(vit), i = 0; !IGRAPH_VIT_END(vit); IGRAPH_VIT_NEXT(vit), i++) {
-            igraph_integer_t node = IGRAPH_VIT_GET(vit);
+            igraph_integer_t v = IGRAPH_VIT_GET(vit);
             igraph_integer_t size, act, edge;
             igraph_vector_int_t *vvec = 0, *evec = 0;
 
@@ -284,20 +284,20 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
             IGRAPH_ALLOW_INTERRUPTION();
 
             size = 0;
-            act = node;
+            act = v;
             while (parent_eids[act]) {
                 size++;
                 edge = parent_eids[act] - 1;
                 act = IGRAPH_OTHER(graph, edge, act);
             }
-            if (vvec && (size > 0 || node == from)) {
+            if (vvec && (size > 0 || v == from)) {
                 IGRAPH_CHECK(igraph_vector_int_resize(vvec, size + 1));
-                VECTOR(*vvec)[size] = node;
+                VECTOR(*vvec)[size] = v;
             }
             if (evec) {
                 IGRAPH_CHECK(igraph_vector_int_resize(evec, size));
             }
-            act = node;
+            act = v;
             while (parent_eids[act]) {
                 edge = parent_eids[act] - 1;
                 act = IGRAPH_OTHER(graph, edge, act);
@@ -334,23 +334,28 @@ igraph_error_t igraph_get_widest_paths(const igraph_t *graph,
  * \ref igraph_get_widest_paths().
  *
  * \param graph The input graph, it can be directed or undirected.
- * \param vertices Pointer to an initialized vector or a null
- *        pointer. If not a null pointer, then the vertex IDs along
- *        the path are stored here, including the source and target
- *        vertices.
- * \param edges Pointer to an initialized vector or a null
- *        pointer. If not a null pointer, then the edge IDs along the
- *        path are stored here.
- * \param from The id of the source vertex.
- * \param to The id of the target vertex.
- * \param weights The edge weights. Edge weights can be negative. If this
- *        is a null pointer or if any edge weight is NaN, then an error
- *        is returned. Edges with positive infinite weight are ignored.
- * \param mode A constant specifying how edge directions are
- *        considered in directed graphs. \c IGRAPH_OUT follows edge
- *        directions, \c IGRAPH_IN follows the opposite directions,
- *        and \c IGRAPH_ALL ignores edge directions. This argument is
- *        ignored for undirected graphs.
+ * \param vertices Pointer to an initialized vector or \c NULL. If not \c NULL,
+ *    then the vertex IDs along the path are stored here, including the source
+ *    and target vertices.
+ * \param edges Pointer to an initialized vector or \c NULL. If not \c NULL,
+ *    then the edge IDs along the path are stored here.
+ * \param from The ID of the source vertex.
+ * \param to The ID of the target vertex.
+ * \param weights The edge weights, interpreted as widths. Edge weights can be
+ *    negative, but must not be NaN. Edges with negative infinite weight are
+ *    ignored. The weight vector is required: if \c NULL is passed, an error is
+ *    raised.
+ * \param mode The type of widest paths to be used for the
+ *        calculation in directed graphs. Possible values:
+ *        \clist
+ *        \cli IGRAPH_OUT
+ *          the outgoing paths are calculated.
+ *        \cli IGRAPH_IN
+ *          the incoming paths are calculated.
+ *        \cli IGRAPH_ALL
+ *          the directed graph is considered as an
+ *          undirected one for the computation.
+ *        \endclist
  * \return Error code.
  *
  * Time complexity: O(|E|log|E|+|V|), |V| is the number of vertices,
@@ -406,12 +411,15 @@ igraph_error_t igraph_get_widest_path(const igraph_t *graph,
  * \brief Widths of widest paths between vertices.
  *
  * This function implements a modified Floyd-Warshall algorithm, to find the
- * widest path widths between a set of source and target vertices. It is
- * primarily useful for all-pairs path widths in very dense graphs, as its
- * running time is manily determined by the vertex count, and is not sensitive
- * to the graph density. In sparse graphs, other methods such as Dijkstra's
- * algorithm, implemented in \ref igraph_widest_path_widths_dijkstra() will
- * perform better.
+ * widest path widths between a set of source and target vertices. The width
+ * of a path is defined as the width of the narrowest edge in the path.
+ *
+ * </para><para>
+ * This algorithm is primarily useful for all-pairs path widths in very dense
+ * graphs, as its running time is manily determined by the vertex count, and
+ * is not sensitive to the graph density. In sparse graphs, other methods such
+ * as Dijkstra's algorithm, implemented in
+ * \ref igraph_widest_path_widths_dijkstra() will perform better.
  *
  * </para><para>
  * Note that internally this function always computes the path width matrix
@@ -428,7 +436,7 @@ igraph_error_t igraph_get_widest_path(const igraph_t *graph,
  * \param from The source vertices.
  * \param to The target vertices.
  * \param weights The edge weights, interpreted as widths. Edge weights can be
- *    negative, but must not be NaN. Edges with positive infinite weight are
+ *    negative, but must not be NaN. Edges with negative infinite weight are
  *    ignored. The weight vector is required: if \c NULL is passed, an error is
  *    raised.
  * \param mode For directed graphs; whether to follow paths along edge
@@ -450,28 +458,28 @@ igraph_error_t igraph_widest_path_widths_floyd_warshall(const igraph_t *graph,
                                    igraph_neimode_t mode) {
 
     /* Implementation Details: This is a modified Floyd Warshall algorithm
-    which computes the widest path between every pair of nodes. The key
+    which computes the widest path between every pair of vertices. The key
     difference between this and the regular Floyd Warshall is that instead
-    of updating the distance between two nodes to be the minimum of itself
-    and the distance through an intermediate node, we instead set the width
-    to be the maximum of itself and the width through the intermediate node.
+    of updating the distance between two vertices to be the minimum of itself
+    and the distance through an intermediate vertex, we instead set the width
+    to be the maximum of itself and the width through the intermediate vertex.
 
-    We denote the widest path from a node to itself as infinity, and the widest
-    path from a node to a node it cannot reach as negative infinity.
+    We denote the widest path from a vertex to itself as infinity, and the widest
+    path from a vertex to a vertex it cannot reach as negative infinity.
     */
 
-    igraph_integer_t no_of_nodes = igraph_vcount(graph);
-    igraph_integer_t no_of_edges = igraph_ecount(graph);
+    const igraph_integer_t vcount = igraph_vcount(graph);
+    const igraph_integer_t ecount = igraph_ecount(graph);
     igraph_bool_t in = false, out = false;
 
     if (! weights) {
         IGRAPH_ERROR("Weight vector is required.", IGRAPH_EINVAL);
     }
 
-    if (igraph_vector_size(weights) != no_of_edges) {
+    if (igraph_vector_size(weights) != ecount) {
         IGRAPH_ERRORF("Weight vector length (%" IGRAPH_PRId ") does not match number of edges (%" IGRAPH_PRId ").",
                       IGRAPH_EINVAL,
-                      igraph_vector_size(weights), no_of_edges);
+                      igraph_vector_size(weights), ecount);
     }
 
     if (igraph_vector_is_any_nan(weights)) {
@@ -497,18 +505,18 @@ igraph_error_t igraph_widest_path_widths_floyd_warshall(const igraph_t *graph,
     }
 
     /* Fill out adjacency matrix */
-    IGRAPH_CHECK(igraph_matrix_resize(res, no_of_nodes, no_of_nodes));
+    IGRAPH_CHECK(igraph_matrix_resize(res, vcount, vcount));
     igraph_matrix_fill(res, -IGRAPH_INFINITY);
-    for (igraph_integer_t i=0; i < no_of_nodes; i++) {
+    for (igraph_integer_t i=0; i < vcount; i++) {
         MATRIX(*res, i, i) = IGRAPH_INFINITY;
     }
 
-    for (igraph_integer_t edge=0; edge < no_of_edges; edge++) {
+    for (igraph_integer_t edge=0; edge < ecount; edge++) {
         igraph_integer_t from = IGRAPH_FROM(graph, edge);
         igraph_integer_t to = IGRAPH_TO(graph, edge);
         igraph_real_t w = VECTOR(*weights)[edge];
 
-        if (w == IGRAPH_INFINITY) {
+        if (w == -IGRAPH_INFINITY) {
             /* Ignore edges with infinite weight */
             continue;
         }
@@ -518,15 +526,15 @@ igraph_error_t igraph_widest_path_widths_floyd_warshall(const igraph_t *graph,
     }
 
     /* Run modified Floyd Warshall */
-    for (igraph_integer_t k = 0; k < no_of_nodes; k++) {
+    for (igraph_integer_t k = 0; k < vcount; k++) {
         /* Iterate in column-major order for better performance */
-        for (igraph_integer_t j = 0; j < no_of_nodes; j++) {
+        for (igraph_integer_t j = 0; j < vcount; j++) {
             igraph_real_t width_kj = MATRIX(*res, k, j);
             if (j == k || width_kj == -IGRAPH_INFINITY) continue;
 
             IGRAPH_ALLOW_INTERRUPTION();
 
-            for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
+            for (igraph_integer_t i = 0; i < vcount; i++) {
                 if (i == j || i == k) continue;
 
                 /* alternative_width := min(A(i,k), A(k,j))
@@ -572,7 +580,7 @@ igraph_error_t igraph_widest_path_widths_floyd_warshall(const igraph_t *graph,
  * \param to The target vertices. It is not allowed to include a vertex twice
  *    or more.
  * \param weights The edge weights, interpreted as widths. Edge weights can be
- *    negative, but must not be NaN. Edges with positive infinite weight are
+ *    negative, but must not be NaN. Edges with negative infinite weight are
  *    ignored. The weight vector is required: if \c NULL is passed, an error is
  *    raised.
  * \param mode For directed graphs; whether to follow paths along edge
@@ -596,21 +604,21 @@ igraph_error_t igraph_widest_path_widths_dijkstra(const igraph_t *graph,
 
     /* Implementation details: This is a Dijkstra algorithm with a
     binary heap, modified to support widest paths. The heap is indexed,
-    so it stores both the widest path to a node, as well as it's index. We
+    so it stores both the widest path to a vertex, as well as it's index. We
     use a 2 way heap so that we can query indexes directly in the heap.
 
     To adapt a Dijkstra to handle widest path, instead of prioritising candidate
-    nodes with the minimum distance, we prioritise those with the maximum
-    width instead. When adding a node into our set of 'completed' nodes, we
-    update all neighbouring nodes with a width that is equal to the min of the
-    width to the current node and the width of the edge.
+    vertices with the minimum distance, we prioritise those with the maximum
+    width instead. When adding a vertex into our set of 'completed' vertices, we
+    update all neighbouring vertices with a width that is equal to the min of the
+    width to the current vertex and the width of the edge.
 
-    We denote the widest path from a node to itself as infinity, and the widest
-    path from a node to a node it cannot reach as negative infinity.
+    We denote the widest path from a vertex to itself as infinity, and the widest
+    path from a vertex to a vertex it cannot reach as negative infinity.
     */
 
-    igraph_integer_t no_of_nodes = igraph_vcount(graph);
-    igraph_integer_t no_of_edges = igraph_ecount(graph);
+    const igraph_integer_t vcount = igraph_vcount(graph);
+    const igraph_integer_t ecount = igraph_ecount(graph);
     igraph_2wheap_t Q;
     igraph_vit_t fromvit, tovit;
     igraph_integer_t no_of_from, no_of_to;
@@ -623,10 +631,10 @@ igraph_error_t igraph_widest_path_widths_dijkstra(const igraph_t *graph,
         IGRAPH_ERROR("Weight vector is required.", IGRAPH_EINVAL);
     }
 
-    if (igraph_vector_size(weights) != no_of_edges) {
+    if (igraph_vector_size(weights) != ecount) {
         IGRAPH_ERRORF("Weight vector length (%" IGRAPH_PRId ") does not match number of edges (%" IGRAPH_PRId ").",
                       IGRAPH_EINVAL,
-                      igraph_vector_size(weights), no_of_edges);
+                      igraph_vector_size(weights), ecount);
     }
 
     if (igraph_vector_is_any_nan(weights)) {
@@ -637,23 +645,23 @@ igraph_error_t igraph_widest_path_widths_dijkstra(const igraph_t *graph,
     IGRAPH_FINALLY(igraph_vit_destroy, &fromvit);
     no_of_from = IGRAPH_VIT_SIZE(fromvit);
 
-    IGRAPH_CHECK(igraph_2wheap_init(&Q, no_of_nodes));
+    IGRAPH_CHECK(igraph_2wheap_init(&Q, vcount));
     IGRAPH_FINALLY(igraph_2wheap_destroy, &Q);
     IGRAPH_CHECK(igraph_lazy_inclist_init(graph, &inclist, mode, IGRAPH_LOOPS));
     IGRAPH_FINALLY(igraph_lazy_inclist_destroy, &inclist);
 
     all_to = igraph_vs_is_all(&to);
     if (all_to) {
-        no_of_to = no_of_nodes;
+        no_of_to = vcount;
     } else {
-        IGRAPH_VECTOR_INT_INIT_FINALLY(&indexv, no_of_nodes);
+        IGRAPH_VECTOR_INT_INIT_FINALLY(&indexv, vcount);
         IGRAPH_CHECK(igraph_vit_create(graph, to, &tovit));
         IGRAPH_FINALLY(igraph_vit_destroy, &tovit);
         no_of_to = IGRAPH_VIT_SIZE(tovit);
         for (i = 0; !IGRAPH_VIT_END(tovit); IGRAPH_VIT_NEXT(tovit)) {
             igraph_integer_t v = IGRAPH_VIT_GET(tovit);
             if (VECTOR(indexv)[v]) {
-                IGRAPH_ERROR("Duplicate vertices in `to', this is not allowed.",
+                IGRAPH_ERROR("Duplicate vertices in target vertex set, this is not allowed.",
                              IGRAPH_EINVAL);
             }
             VECTOR(indexv)[v] = ++i;
@@ -705,8 +713,8 @@ igraph_error_t igraph_widest_path_widths_dijkstra(const igraph_t *graph,
                 igraph_bool_t active = igraph_2wheap_has_active(&Q, tto);
                 igraph_bool_t has = igraph_2wheap_has_elem(&Q, tto);
                 igraph_real_t curwidth = active ? igraph_2wheap_get(&Q, tto) : IGRAPH_INFINITY;
-                if (edgewidth == IGRAPH_INFINITY) {
-                    /* Ignore edges with infinite weight */
+                if (edgewidth == -IGRAPH_INFINITY) {
+                    /* Ignore edges with negative infinite weight */
                 } else if (!has) {
                     /* This is the first time assigning a width to this vertex */
                     IGRAPH_CHECK(igraph_2wheap_push_with_index(&Q, tto, altwidth));